requirements.txt
pandas
numpy
psycopg2-binary
python-dotenv
openpyxl==3.0.10
XlsxWriter
fastapi
uvicorn
requests
hypercorn
